+++ Sora/PeerChannel.swift
@@ -294,10 +294,6 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
                      message: "did connect to signaling channel")
 
         var role: SignalingRole
-        var multistream = configuration.multistreamEnabled
-        if configuration.spotlightEnabled == .enabled {
-            multistream = true
-        }
         switch configuration.role {
         case .sendonly:
             role = .sendonly
@@ -305,11 +301,9 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
             role = .recvonly
         case .sendrecv:
             role = .sendrecv
-            multistream = true
         }
 
         let soraClient = "Sora iOS SDK \(SDKInfo.version)"
-
         let webRTCVersion = "Shiguredo-build \(WebRTCInfo.version) (\(WebRTCInfo.version.dropFirst()).\(WebRTCInfo.branch).\(WebRTCInfo.commitPosition).\(WebRTCInfo.maintenanceVersion) \(WebRTCInfo.shortRevision))"
 
         let simulcast = configuration.simulcastEnabled
@@ -321,7 +315,7 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
             metadata: configuration.signalingConnectMetadata,
             notifyMetadata: configuration.signalingConnectNotifyMetadata,
             sdp: sdp,
-            multistreamEnabled: multistream,
+            multistreamEnabled: configuration.multistreamEnabled,
             videoEnabled: configuration.videoEnabled,
             videoCodec: configuration.videoCodec,
             videoBitRate: configuration.videoBitRate,