+++ CHANGES.md
@@ -11,6 +11,12 @@
 
 ## develop
 
+- [CHANGE] connect メッセージの `multistream` を true 固定で送信する処理を削除する破壊的変更
+  - Configration.role に .sendrecv を指定している場合に multistream を true に更新する処理を削除
+  - Configration.spotlightEnabled に .enabled を指定している場合に multistream を true に更新する処理を削除
+  - 結果、connect メッセージには Configration.multistreamEnabled に指定した値が送信される
+  - 今後は Configration.role に .sendrecv を指定している場合または Configration.spotlightEnabled に .enabled を指定している場合に Confgration.multistreamEnabled に false を指定すると接続エラーになる 
+  - @zztkm
 - [UPDATE] multistreamEnabled を非推奨扱いにする
   - multistreamEnabled の設定が不要なイニシャライザを `Configuration` に追加する
   - ドキュメントコメントに非推奨扱いの旨を追加する